+++ hosts/darwin/codex/default.nix
@@ -14,6 +14,7 @@ in
   imports = [
     ../common-darwin.nix
     ../../../modules/darwin/certbot.nix
+    ../../../modules/darwin/syncthing-password.nix
     ../../../modules/common/certbot-secrets.nix
   ];